+++ assets/javascripts/kitten/index.js
@@ -8,6 +8,10 @@ export { VisuallyHidden } from './components/accessibility/visually-hidden'
 // Accordeon
 export { Accordeon } from './components/accordeon'
 
+// Atoms
+export { Loader } from './components/atoms/loader'
+export { LoaderWithParagraph } from './components/atoms/loader-with-paragraph'
+
 // Avatar
 export { AvatarWithTextAndBadge } from './components/avatar/avatar-with-text-and-badge'
 
@@ -164,10 +168,6 @@ export { CheckList } from './components/lists/check-list'
 export { NavList } from './components/lists/nav-list'
 export { TagList } from './components/lists/tag-list'
 export { Timeline } from './components/lists/timeline'
-
-// Loaders
-export { LoaderWithParagraph } from './components/loaders/loader-with-paragraph'
-export { Loader } from './components/loaders/loader'
 
 // Menus
 export { HeaderMenu } from './components/menus/header-menu'